requirements.txt
python-docx>=2.8.1
openai>=1.12.0
python-dotenv>=1.0.0
difflib 